To Nix, or not to Nix - that is the question
I think the idea of Nix is a cool one. It would be awesome to have a way
of ensuring a consistent development environment wherever I will happen
to work on the project. Nix seems a bit arcane at the moment, though...

Anyway, let's explore.

What I would like to have is a script to build and test the VM in a
nix-shell--with both GCC and Clang--that would be independent of the
machine on which it is running. This should be doable, and provide a
really convenient process:

    ]$ git clone .../viuavm
    ]$ cd viuavm
    # This would build the VM's toolchain using gcc.
    ]$ bash ./nix-build.sh gcc
    # This would test the result.
    ]$ bash ./nix-test.sh

No messing around with dependencies if you only want to build, test, and
work on the code. You would only have to provide the dependencies if you
wanted to install the VM on your machine. The only things that would be
needed to work on the VM without would be Git, Nix, and BASH.